backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
python-dotenv
passlib[bcrypt]
python-jose
bcrypt
email-validator
pydantic
pydantic-settings
alembic
groq